+++ src/components/calendar/month/CalendarMonthly.vue
@@ -24,7 +24,7 @@
             <td class="gts-print-calendar-monthly-content-data" v-for="(day, index ) in calendarData.weekDays" :key="index" @click="selectDay(week[index])"
               :class="getContentCssClass[weekIndex][index]">
 
-              <CalendarMonthlyInfo :calendarDay="week[index]" :selectedDay="selectedDay"/>
+              <CalendarMonthlyInfo :calendarDay="week[index]" :selectedDay="selectedDay" :contextMenuActions="contextMenuActions"/>
 
             </td> 
         
@@ -42,11 +42,13 @@
 
 <script>
 
-import CalendarMonthlyInfo from "./CalendarMonthlyInfo";
+import CalendarMonthlyInfo from "./CalendarMonthlyInfo.vue";
 
 export default {
 
   name: "CalendarMonthly",
+
+  emits : ['daySelected'],
 
   components: {
     CalendarMonthlyInfo
@@ -57,6 +59,12 @@ export default {
       type: Object,
       required: true,
     },
+    
+    contextMenuActions: {
+			type: Array,
+			required: false,
+			default: () => [],
+		},
   },
 
   data () {
@@ -72,7 +80,7 @@ export default {
           if (this.selectedDay === day) {
             return "selected";
           }
-          if (day.timeSlots.length === 0) {
+          if (day.disabled) {
             return "light-gray";
           }
           return "";
@@ -84,8 +92,11 @@ export default {
   methods : {
 
     selectDay(day){
-      if(day.timeSlots.length>0)
+      if(!day.disabled){
         this.selectedDay = day;
+        this.$emit("daySelected", this.selectedDay);
+      }
+       
     },
 
     displayMenu(day){
@@ -105,7 +116,6 @@ export default {
   .gts-print-calendar-monthly-container {
     border-radius: 12px 12px 0 0;
     border: 1px solid $neutral-color-200;
-    overflow: hidden;
   
     .gts-print-calendar-monthly{
       border-collapse: collapse;